chore: rename components/TooltipTrigger folder to components/PopupTooltipTrigger
